# etcdserver: add trace for txn request