--- a/pkg/traceutil/trace.go
+++ b/pkg/traceutil/trace.go
@@ -60,12 +60,14 @@ type Trace struct {
 	startTime    time.Time
 	steps        []step
 	stepDisabled bool
+	isEmpty      bool
 }
 
 type step struct {
-	time   time.Time
-	msg    string
-	fields []Field
+	time       time.Time
+	msg        string
+	fields     []Field
+	isSubTrace bool
 }
 
 func New(op string, lg *zap.Logger, fields ...Field) *Trace {
@@ -74,7 +76,7 @@ func New(op string, lg *zap.Logger, fields ...Field) *Trace {
 
 // TODO returns a non-nil, empty Trace
 func TODO() *Trace {
-	return &Trace{}
+	return &Trace{isEmpty: true}
 }
 
 func Get(ctx context.Context) *Trace {
@@ -93,7 +95,7 @@ func (t *Trace) SetStartTime(time time.Time) {
 }
 
 func (t *Trace) InsertStep(at int, time time.Time, msg string, fields ...Field) {
-	newStep := step{time, msg, fields}
+	newStep := step{time: time, msg: msg, fields: fields}
 	if at < len(t.steps) {
 		t.steps = append(t.steps[:at+1], t.steps[at:]...)
 		t.steps[at] = newStep
@@ -102,11 +104,24 @@ func (t *Trace) InsertStep(at int, time time.Time, msg string, fields ...Field) 
 	}
 }
 
+// SubTrace adds step to trace as a sublevel trace
+func (t *Trace) SubTrace(fields ...Field) {
+	t.steps = append(t.steps, step{fields: fields, isSubTrace: true})
+}
+
 // Step adds step to trace
 func (t *Trace) Step(msg string, fields ...Field) {
 	if !t.stepDisabled {
 		t.steps = append(t.steps, step{time: time.Now(), msg: msg, fields: fields})
 	}
+}
+
+// StepWithFunction will measure the input function as a single step
+func (t *Trace) StepWithFunction(f func(), msg string, fields ...Field) {
+	t.DisableStep()
+	f()
+	t.EnableStep()
+	t.Step(msg, fields...)
 }
 
 // DisableStep sets the flag to prevent the trace from adding steps
@@ -121,8 +136,14 @@ func (t *Trace) EnableStep() {
 
 func (t *Trace) AddField(fields ...Field) {
 	for _, f := range fields {
-		t.fields = append(t.fields, f)
+		if !t.updateFieldIfExist(f) {
+			t.fields = append(t.fields, f)
+		}
 	}
+}
+
+func (t *Trace) IsEmpty() bool {
+	return t.isEmpty
 }
 
 // Log dumps all steps in the Trace
@@ -154,7 +175,14 @@ func (t *Trace) logInfo(threshold time.Duration) (string, []zap.Field) {
 
 	var steps []string
 	lastStepTime := t.startTime
-	for _, step := range t.steps {
+	for i := 0; i < len(t.steps); i++ {
+		step := t.steps[i]
+		if step.isSubTrace {
+			for j := i + 1; j < len(t.steps) && !t.steps[j].isSubTrace; j++ {
+				t.steps[j].fields = append(step.fields, t.steps[j].fields...)
+			}
+			continue
+		}
 		stepDuration := step.time.Sub(lastStepTime)
 		if stepDuration > threshold {
 			steps = append(steps, fmt.Sprintf("trace[%d] '%v' %s (duration: %v)",
@@ -167,6 +195,17 @@ func (t *Trace) logInfo(threshold time.Duration) (string, []zap.Field) {
 		zap.Duration("duration", totalDuration),
 		zap.Time("start", t.startTime),
 		zap.Time("end", endTime),
-		zap.Strings("steps", steps)}
+		zap.Strings("steps", steps),
+		zap.Int("step_count", len(t.steps))}
 	return msg, fs
+}
+
+func (t *Trace) updateFieldIfExist(f Field) bool {
+	for i, v := range t.fields {
+		if v.Key == f.Key {
+			t.fields[i].Value = f.Value
+			return true
+		}
+	}
+	return false
 }