Here is an example output of a txn request. 
Request:
```
cmp := clientv3.Compare(clientv3.Value("foo"), "=", "bar");
op1 := clientv3.OpGet("foo", clientv3.WithRange("fooo"))
op2 := clientv3.OpPut("txn", "put")
op3 := clientv3.OpPut("put","txn")
op4 := clientv3.OpTxn([]clientv3.Cmp{cmp}, []clientv3.Op{op1}, []clientv3.Op{})
txn := cli.Txn(ctx).If(cmp).Then(op1, op2, op3, op4)
```
Output:
```
{"level":"info","ts":"2020-01-02T15:27:52.215-0800","caller":"traceutil/trace.go:166","msg":"trace[1023984242] transaction",
"detail":"{read_only:false; response_revision:7; number_of_response:4; }","duration":"440.713µs","start":"2020-01-02T15:27:52.215-0800","end":"2020-01-02T15:27:52.215-0800",
"steps":["trace[1023984242] 'process raft request'  (duration: 195.497µs)",
"trace[1023984242] 'compare'  (duration: 36.258µs)","trace[1023984242] 'check requests'  (duration: 1.317µs)",
"trace[1023984242] 'range keys from in-memory index tree' {req_type:range; range_begin:foo; range_end:fooo; } (duration: 12.56µs)",
"trace[1023984242] 'range keys from bolt db' {req_type:range; range_begin:foo; range_end:fooo; } (duration: 2.126µs)",
"trace[1023984242] 'filter and sort the key-value pairs' {req_type:range; range_begin:foo; range_end:fooo; } (duration: 1.034µs)",
"trace[1023984242] 'assemble the response' {req_type:range; range_begin:foo; range_end:fooo; } (duration: 308ns)",
"trace[1023984242] 'get key's previous created_revision and leaseID' {req_type:put; key:txn; req_size:10; } (duration: 9.659µs)",
"trace[1023984242] 'marshal mvccpb.KeyValue' {req_type:put; key:txn; req_size:10; } (duration: 1.766µs)",
"trace[1023984242] 'store kv pair into bolt db' {req_type:put; key:txn; req_size:10; } (duration: 9.589µs)",
"trace[1023984242] 'attach lease to kv pair' {req_type:put; key:txn; req_size:10; } (duration: 61ns)",
"trace[1023984242] 'get key's previous created_revision and leaseID' {req_type:put; key:put; req_size:10; } (duration: 1.192µs)",
"trace[1023984242] 'marshal mvccpb.KeyValue' {req_type:put; key:put; req_size:10; } (duration: 306ns)",
"trace[1023984242] 'store kv pair into bolt db' {req_type:put; key:put; req_size:10; } (duration: 1.139µs)",
"trace[1023984242] 'attach lease to kv pair' {req_type:put; key:put; req_size:10; } (duration: 47ns)",
"trace[1023984242] 'range keys from in-memory index tree' {req_type:range; range_begin:foo; range_end:fooo; } (duration: 2.293µs)",
"trace[1023984242] 'range keys from bolt db' {req_type:range; range_begin:foo; range_end:fooo; } (duration: 897ns)",
"trace[1023984242] 'filter and sort the key-value pairs' {req_type:range; range_begin:foo; range_end:fooo; } (duration: 146ns)",
"trace[1023984242] 'assemble the response' {req_type:range; range_begin:foo; range_end:fooo; } (duration: 112ns)"],
"step_count":19}
```
This example sets the `traceThreshold ` to 0 to show every step of the request. Normally, the step will be logged out only when the duration exceeds 100ms.
https://github.com/etcd-io/etcd/blob/e6980b1f9fc008837abb9177e87e893e48d565c1/etcdserver/v3_server.go#L42

Here are the benchmark results from 10000 times run of the above example which will trigger trace logging.
Without trace:
![image](https://user-images.githubusercontent.com/11894762/71855697-9fa77180-3096-11ea-9306-faa5f99ee339.png)
With trace:
![image](https://user-images.githubusercontent.com/11894762/71855704-a635e900-3096-11ea-9905-b12f23e22868.png)