@xiang90 2020-01-04
how much would this affect performance?