@YoyinZyc 2020-01-06
> how much would this affect performance?

Added benchmark results above.